add: rename ccg files
