+++ src/ccg/mod.rs
@@ -1,6 +1,6 @@
-pub mod language_parser;
-mod ccg_type;
-mod ccg_rule;
-pub mod ccg_node;
-mod ccg_word;
-pub mod ccg_type_parser;
+pub mod sentence_parser;
+mod r#type;
+mod rule;
+pub mod node;
+mod word;
+pub mod type_parser;
