+++ src/ccg/type.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 use serde::{Deserialize, Deserializer};
-use crate::ccg::ccg_type_parser::parse_category;
+use crate::ccg::type_parser::parse_category;
 #[derive(Debug, Clone, PartialEq)]
 pub enum CCGType {
     ForwardsFunctor(Box<CCGType>, Box<CCGType>),
